+++ src/dns_exporter/config.py
@@ -266,7 +266,7 @@ class Config:
 
         # validate proxy
         if self.proxy:
-            # proxy support doesn't work for DoT and DoQ for now
+            # proxy support doesn't work for DoT for now
             if self.protocol in ["dot"]:
                 logger.error(f"proxy not valid for protocol {self.protocol}")
                 raise ConfigError(